pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lis-personal-assistant"
version = "0.1.0"
description = "A lean, modular knowledge management and workflow system for programme management"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "typer[all]>=0.15",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "python-frontmatter>=1.1",
    "openai>=1.60",
    "python-dotenv>=1.0",
    "python-multipart>=0.0.18",
]

[project.optional-dependencies]
notion = ["notion-client>=2.2"]
publish = ["mkdocs-material>=9.5"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.28",
    "ruff>=0.8",
]

[project.scripts]
assistant = "assistant.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/assistant"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
